+++ asforepa/src/test/java/de/gematik/idp/asforepa/controller/AuthorizationControllerTest.java
@@ -37,9 +37,9 @@ import java.security.interfaces.RSAPrivateKey;
 import java.time.ZonedDateTime;
 import java.util.Base64;
 import java.util.function.UnaryOperator;
-import kong.unirest.HttpResponse;
-import kong.unirest.JsonNode;
-import kong.unirest.Unirest;
+import kong.unirest.core.HttpResponse;
+import kong.unirest.core.JsonNode;
+import kong.unirest.core.Unirest;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.jose4j.jws.AlgorithmIdentifiers;
@@ -64,10 +64,14 @@ class AuthorizationControllerTest {
   private String testHostUrl;
   private JsonWebToken clientAttest;
   private JsonWebToken invalidClientAttest;
+  private JsonWebToken invalidClientAttestBase64;
+  private String invalidAuthCodeBase64;
 
   private static final String VALID_USER_AGENT = "CLIENTID1234567890AB/2.1.12-45";
   private static final String VALID_AUTH_CODE =
       "eyJhbGciOiJkaXIiLCJlbmMiOiJBMjU2R0NNIiwiY3R5IjoiTkpXVCIsImV4cCI6MTcwMTM0MTkxNX0..YokE-g-EdrYxp6BK.CNldvZZwQuPHfkl3X9dNXVjk2M2LyKj_3A85dtEOGfAG5knjl7Q9P5ce8WoVp8SiXmNm63eUI9XcpFOjdAjxVTrHufnRUYjMZY4VZvXhqDW1Zalz_qC7eVNCiAZE2nXy6ozeLJmibLhgp2flLCGT-Ap1sVH8u6LZflcu-cIPhYR89A2pUh40Kg0ItCtJ1dqG6vinlsMRLs8t2oc5G4-gmI6O-1IlN2ekSTS6zGkq301YueHY8xGyij9SPoIxoiwBuo5C2lDBjWXNMCTKy9JPEl4S3vevg9UFO4bGaw5myoH8xN-S07ZKm3EnkvlzKXdTrBmcFusKE9NOBH4fgLmO4AFHqCEVDmHm7OAejVpRueSKAQZ18VZFUkqPdBYjFkpI_-Q45qBVIVAICsXFSa62LO6uZw6qDeME7c4NonTJCijcQ-RvFGc5Am2A7uJ1jzxiocpU4qRume3V-yWn9_tz0gcBqfUa2ejM2SziXg3PQzYYJ7bxTzWvbuNtBQhA_wzxr8eWf-4Z3NZSsuGzkX3ru5xTDrAJvivm01MqySUZkJz4Ho-kwJ2Fef5sVVMx8EN5fdxvYKUpGtco214a5gdFwVmPg3IiroXR264KWRP9lMgkBLgCyeQXQ07dR4_vl9uU47ytVFzQdshVpK3-1kSq_4SFEzvixCLZqRR4Mv7-PjF8Jmcdftp1jt7zg0Syt7PLYY4hCJCWe_Ftk2G7QD79kPPkvxwKbP1MqxWwliN6uUIYN8UMaIvu-6oPLYeoa8BaAoQiG9tdFL_UXXAGB-tW-VT-1ONofwr6_ZJI7n4jtO5-AZ1ccS1Oocqsc9kDnsFfNouMPTp0HcS690LN1oP-RkRnA3c-dmdSCJdsjsrI2I0tkh5pxlWs_Sg_vn10BjGtYaQHLGdT8cEf8nSRtrZLj1HkAYc5uxuVzJ84enIwQkFEn6dwJRShglxWw9DCWv7sTvww_PNw1Kt8BrzXuPJ7pOP5qi2MJjkOaJ-gqIp4NzGQ7n4Q1vv4ERgTiZlLSm_7fofd_GN4QFp-45DpnHXPtKyAKDbVXTh73myiPbgIIlhG7aaO6PW4aw7z2VuPaVXhv0932UdkdQ7CvJrWDsnLmUIviu72Z7uFA7aI4ilpT4yPcdv2c3Se1cnG4mITOGycBfMtX41tglv5k-YjMdzocWegKgZKwk6hk39O26FxLwto_xfr_2U2_y1S67dRviCRUcCPSmusYUtxKtb_mG-fGyt6hrlWN26y5LKvmB0mUH6kyUSTzyfzLmc4M6ExCs2cO8JpSGENHP5itwMCQ-HxRJ1sfH2LlMn7ECSaz6kPeaKPT2Rrvxck0GF2R-dqY_NMTeC8CV9BJWP9HNpGbxELe7j_RBkPcwXednfHFRBd8r24rZn7gHVi2Dd33g.Xtx_4AO2Gfbz1NeyzfSmaw";
+  private static final String AUTH_CODE_BASE64 =
+      "eyJhbGciOiJkaXIiLCJlbmMiOiJBMjU2R0NNIiwiY3R5IjoiTkpXVCIsImV4cCI6MTcwMTM0MTkxNX0..YokE+g+EdrYxp6BK.CNldvZZwQuPHfkl3X9dNXVjk2M2LyKj/3A85dtEOGfAG5knjl7Q9P5ce8WoVp8SiXmNm63eUI9XcpFOjdAjxVTrHufnRUYjMZY4VZvXhqDW1Zalz/qC7eVNCiAZE2nXy6ozeLJmibLhgp2flLCGT+Ap1sVH8u6LZflcu+cIPhYR89A2pUh40Kg0ItCtJ1dqG6vinlsMRLs8t2oc5G4+gmI6O+1IlN2ekSTS6zGkq301YueHY8xGyij9SPoIxoiwBuo5C2lDBjWXNMCTKy9JPEl4S3vevg9UFO4bGaw5myoH8xN+S07ZKm3EnkvlzKXdTrBmcFusKE9NOBH4fgLmO4AFHqCEVDmHm7OAejVpRueSKAQZ18VZFUkqPdBYjFkpI/+Q45qBVIVAICsXFSa62LO6uZw6qDeME7c4NonTJCijcQ+RvFGc5Am2A7uJ1jzxiocpU4qRume3V+yWn9/tz0gcBqfUa2ejM2SziXg3PQzYYJ7bxTzWvbuNtBQhA/wzxr8eWf+4Z3NZSsuGzkX3ru5xTDrAJvivm01MqySUZkJz4Ho+kwJ2Fef5sVVMx8EN5fdxvYKUpGtco214a5gdFwVmPg3IiroXR264KWRP9lMgkBLgCyeQXQ07dR4/vl9uU47ytVFzQdshVpK3+1kSq/4SFEzvixCLZqRR4Mv7+PjF8Jmcdftp1jt7zg0Syt7PLYY4hCJCWe/Ftk2G7QD79kPPkvxwKbP1MqxWwliN6uUIYN8UMaIvu+6oPLYeoa8BaAoQiG9tdFL/UXXAGB+tW+VT+1ONofwr6/ZJI7n4jtO5+AZ1ccS1Oocqsc9kDnsFfNouMPTp0HcS690LN1oP+RkRnA3c+dmdSCJdsjsrI2I0tkh5pxlWs/Sg/vn10BjGtYaQHLGdT8cEf8nSRtrZLj1HkAYc5uxuVzJ84enIwQkFEn6dwJRShglxWw9DCWv7sTvww/PNw1Kt8BrzXuPJ7pOP5qi2MJjkOaJ+gqIp4NzGQ7n4Q1vv4ERgTiZlLSm/7fofd/GN4QFp+45DpnHXPtKyAKDbVXTh73myiPbgIIlhG7aaO6PW4aw7z2VuPaVXhv0932UdkdQ7CvJrWDsnLmUIviu72Z7uFA7aI4ilpT4yPcdv2c3Se1cnG4mITOGycBfMtX41tglv5k+YjMdzocWegKgZKwk6hk39O26FxLwto/xfr/2U2/y1S67dRviCRUcCPSmusYUtxKtb/mG+fGyt6hrlWN26y5LKvmB0mUH6kyUSTzyfzLmc4M6ExCs2cO8JpSGENHP5itwMCQ+HxRJ1sfH2LlMn7ECSaz6kPeaKPT2Rrvxck0GF2R+dqY/NMTeC8CV9BJWP9HNpGbxELe7j/RBkPcwXednfHFRBd8r24rZn7gHVi2Dd33g.Xtx/4AO2Gfbz1NeyzfSmaw";
 
   @SneakyThrows
   @BeforeAll
@@ -157,6 +161,42 @@ class AuthorizationControllerTest {
                         }));
 
     clientAttest = new JsonWebToken(signedJwt);
+
+    final String nonceB64 =
+        Unirest.get(testHostUrl + NONCE_ENDPOINT)
+            .header(X_USERAGENT, VALID_USER_AGENT)
+            .asJson()
+            .getBody()
+            .getObject()
+            .getString("nonce");
+
+    final JwtClaims claimsB64 = new JwtClaims();
+    claims.setClaim(ClaimName.NONCE.getJoseName(), nonceB64);
+    claims.setClaim(ClaimName.ISSUED_AT.getJoseName(), ZonedDateTime.now().toEpochSecond());
+    final JsonWebSignature jwsBase64 = new JsonWebSignature();
+    jwsBase64.setPayload(claimsB64.toJson());
+    jwsBase64.setHeader("typ", "JWT");
+    jwsBase64.setCertificateChainHeaderValue(smcbIdentityEcc.getCertificate());
+    jwsBase64.setAlgorithmHeaderValue(AlgorithmIdentifiers.ECDSA_USING_P256_CURVE_AND_SHA256);
+    jwsBase64.setKey(smcbIdentityEcc.getPrivateKey());
+    final String signedJwtBase64 =
+        jwsBase64.getHeaders().getEncodedHeader()
+            + "."
+            + jwsBase64.getEncodedPayload()
+            + "."
+            + Base64.getEncoder()
+                .encodeToString(
+                    getSignatureBytes(
+                        getContentSigner(smcbIdentityEcc),
+                        jwsBase64,
+                        sigData -> {
+                          try {
+                            return convertDerToConcatenated(sigData, 64);
+                          } catch (final IOException e) {
+                            throw new AsEpaException(e);
+                          }
+                        }));
+    invalidClientAttestBase64 = new JsonWebToken(signedJwtBase64);
   }
 
   private byte[] getSignatureBytes(
@@ -300,14 +340,16 @@ class AuthorizationControllerTest {
   void validSendAuthCodeRequest_response200_checkVNPisHex() {
     final AuthCodeRequest authCodeRequest =
         new AuthCodeRequest(VALID_AUTH_CODE, clientAttest.getRawString());
-    assertThat(
-            Unirest.post(testHostUrl + AUTH_CODE_ENDPOINT)
-                .header("Content-Type", "application/json")
-                .header(X_USERAGENT, VALID_USER_AGENT)
-                .body(authCodeRequest)
-                .asString()
-                .getBody())
-        .isHexadecimal();
+
+    final JsonNode authorizationResponse =
+        Unirest.post(testHostUrl + AUTH_CODE_ENDPOINT)
+            .header("Content-Type", "application/json")
+            .header(X_USERAGENT, VALID_USER_AGENT)
+            .body(authCodeRequest)
+            .asJson()
+            .getBody();
+    assertThat(authorizationResponse.getObject().keySet()).containsExactlyInAnyOrder("vau-np");
+    assertThat(authorizationResponse.getObject().get("vau-np").toString()).isHexadecimal();
   }
 
   @Test
@@ -384,6 +426,66 @@ class AuthorizationControllerTest {
                 .getObject()
                 .getString("error_description"))
         .contains("auth code is invalid");
+  }
+
+  @Test
+  void invalidSendAuthCodeRequest_base64AuthCode() {
+    final AuthCodeRequest authCodeRequest =
+        new AuthCodeRequest(AUTH_CODE_BASE64, clientAttest.getRawString());
+    assertThat(
+            Unirest.post(testHostUrl + AUTH_CODE_ENDPOINT)
+                .header("Content-Type", "application/json")
+                .header(X_USERAGENT, VALID_USER_AGENT)
+                .body(authCodeRequest)
+                .asString()
+                .getStatus())
+        .isEqualTo(HttpStatus.BAD_REQUEST.value());
+  }
+
+  @Test
+  void invalidSendAuthCodeRequest_base64AuthCode_message() {
+    final AuthCodeRequest authCodeRequest =
+        new AuthCodeRequest(AUTH_CODE_BASE64, clientAttest.getRawString());
+    assertThat(
+            Unirest.post(testHostUrl + AUTH_CODE_ENDPOINT)
+                .header("Content-Type", "application/json")
+                .header(X_USERAGENT, VALID_USER_AGENT)
+                .body(authCodeRequest)
+                .asJson()
+                .getBody()
+                .getObject()
+                .getString("error_description"))
+        .contains("auth code doesn't match base64url pattern");
+  }
+
+  @Test
+  void invalidSendAuthCodeRequest_base64ClientAttest() {
+    final AuthCodeRequest authCodeRequest =
+        new AuthCodeRequest(VALID_AUTH_CODE, invalidClientAttestBase64.getRawString());
+    assertThat(
+            Unirest.post(testHostUrl + AUTH_CODE_ENDPOINT)
+                .header("Content-Type", "application/json")
+                .header(X_USERAGENT, VALID_USER_AGENT)
+                .body(authCodeRequest)
+                .asString()
+                .getStatus())
+        .isEqualTo(HttpStatus.BAD_REQUEST.value());
+  }
+
+  @Test
+  void invalidSendAuthCodeRequest_base64ClientAttest_message() {
+    final AuthCodeRequest authCodeRequest =
+        new AuthCodeRequest(VALID_AUTH_CODE, invalidClientAttestBase64.getRawString());
+    assertThat(
+            Unirest.post(testHostUrl + AUTH_CODE_ENDPOINT)
+                .header("Content-Type", "application/json")
+                .header(X_USERAGENT, VALID_USER_AGENT)
+                .body(authCodeRequest)
+                .asJson()
+                .getBody()
+                .getObject()
+                .getString("error_description"))
+        .contains("client attest doesn't match base64url pattern");
   }
 
   @Test